Extract interface from RealizableTaskContainer
